Support multiple VideoPlayers, + template debug
This commit refactors the existing VideoPlayer template and JS so that
multiple VideoPlayers can be included on one page. It adds a template
debug view for the VideoPlayer module for easier testing and debugging.
It also simplifies the interface to the VideoPlayer so that users only
have to provide a YouTube ID instead of both an ID and a URL.

New template debug view:
http://localhost:8000/admin/template_debug/v1/video_player/