Ensure that translations are always deactivated
Adds DeactivateTranslationsMiddleware to ensure that Django translations
are always deactivated at the end of a web request. If we don't do this,
the activated translation will hang around on the thread local, and
could be used on subsequent requests.

As a test example with the local runserver, first visit this page, and
note header and footer are properly in English:

http://localhost:8000/rural-or-underserved-tool/

Then visit this page, which activates Spanish translations:

http://localhost:8000/consumer-tools/financial-well-being/es/

Note the header and footer are (properly) in Spanish there. Next visit
the first page again:

http://localhost:8000/rural-or-underserved-tool/

The header and footer get stuck in Spanish!

This commit fixes this by always deactivating translations.