Deprecate Wagtail welcome page
This page can currently only be seen by those users with Django admin
permission (developers and admins). Those users don't always get sent to the
page upon logging in, but only when logging in with certain ?next=/something/
parameters in the login URL.

Besides that, the content and design of this page has some issues, so
we've decided to remove it. This PR does so.

See platform#3800 for background.