Use a test index for Haystack in tests
When our tests run, Haystack uses the Elasticsearch index configured in `settings/base.py`. Locally, this can can have data in it that we've placed there either with `update_index`, or in some other way. This can interfere with tests.

This change adds a `test_` prefix to the index name in our test-specific settings.