Fix: FCM heading and body should be required
Per the design system:

https://cfpb.github.io/design-system/patterns/featured-content-module

Featured Content modules should always have a heading and a body. These
fields should thus be required in Wagtail. The template for this module
already assumes that these fields are defined, so this just corrects how
validation works in the admin.

I've audited the existing FCMs and confirm that none have empty headings
or bodies:

https://nbviewer.jupyter.org/gist/chosak/5ed2df7ad234a88ec36a64ffa02e4418

This commit also fixes the design manual URL referenced in the FCM
template.