Turn on Wagtail 2.5's Postgres full-text search
The upgrade to 2.5 removed an obstacle to indexed admin searching
because it removed the naming clash between Wagtail and Haystack over
the `upndate_index`c command.

This version also makes available Postgres' full-text search as an alternative
to Elasticsearch, and in intiail testing it works well. It has the added
advantage of simplicity.

Our current vanilla admin search is not recommended for use in production.
The Elasticsearch backend is recommended for sites with more than a million
pages, and we have 8,741. The indexing picks up snippets and metadata for
documents and images as well as page content.

Testing
To give it a try, check out the branch, migrate and index:

```
./cfgov/manage.py migrate
./cfgov/manage.py wagtail_update_index
```

The try some searches in the admin search pane.