Add a 'validate-migrations' tox env
This change removes our custom env-var based migrations check that only occurs in CI and moves it to a separate tox env that can be invoked along side linting and unit testing. It also runs the migrations check only against the production settings, so it only catches apps that we use in production.