Remove the comparisontool URL include
In order to allow Wagtail take over serving the paying-for-college
landing page, we need to remove the redirect URL that sent that root
node to comparisontool for handling.

We need to retain three downstream comparisontool pages for now, so
this patch adds explicit URL entries for those pages to cfgov.urls.py.

Testing
After checkint out the branch, make sure these pages render properly:
- http://localhost:8000/paying-for-college/ should show the new wagtail landing page:

[screen shot]

These three pages should be unchanged, and show a header with pencils:
- http://localhost:8000/paying-for-college/repay-student-debt/
- http://localhost:8000/paying-for-college/choose-a-student-loan/
- http://localhost:8000/paying-for-college/manage-your-college-money/

[screen shot]