Remove duplicate OAH social media template
The BAH code includes its own social media template which is mostly a
duplicate of the generic one. The BAH one is only used on one specific
page, which is the Explore Rates page [0].

This commit removes the custom module in favor of using the generic one.
In practice this makes no difference. The generic one happens to include
an (optional) print option, that is not used on this specific page.

This change will make it easier to add proper Spanish support to the
social media molecule.

[0] https://www.consumerfinance.gov/owning-a-home/explore-rates/