Link Spanish pages to Spanish social media
When used in its non-sharing configuration, our social media molecule
links to the Bureau's English Facebook (cfpb) and Twitter (cfpb)
accounts. On Spanish pages, these links should go to the appropriate
Spanish Facebook (CFPBespanol) and Twitter (CFPBespanol) accounts.

We always import this social-media template `with context` so it should
have the proper `language` variable set.

There's some additional potential work to be done here when the social
media molecule is used for sharing, for example around the default text,
but that is not covered here.

To test, run a local server and visit these pages, and observe the links
in the footer. These should be English:

- http://localhost:8000/
- http://localhost:8000/about-us/blog/receive-your-unemployment-benefits-options/

And these will be Spanish:

- http://localhost:8000/es/
- http://localhost:8000/about-us/blog/opciones-recibir-beneficios-desempleo/