Bump Jinja2 version from 2.10.1 to 2.11.2
This change bumps the version of the Jinja2 package we use from 2.10.2
to 2.11.2, released on 4/13/2020.

See changelog here:
https://jinja.palletsprojects.com/en/2.11.x/changelog/

There are various changes; the one in particular motivating me to make
this change is the addition of the {% debug %} tag which makes it easier
to debug template context. Use of this tag prints out a bunch of
template debugger context as a JSON blob:

https://jinja.palletsprojects.com/en/2.11.x/extensions/#debug-extension

Because we don't want to accidentally do this in production, I've set
this up so that this extension only gets turned on when DEBUG is True.
See the code comment for a little more detail.

Another nice change in this release is greater flexibility around using
logic like

{{ 'foo' if bar }}

instead of

{{ 'foo' if  bar else '' }}

All tests pass with this change.