Deprecate signed_redirect, unsigned_redirect tags
This commit deprecates the definition and use of two template tags:
signed_redirect and unsigned_redirect. These template tags are used to
modify external linked URLs so that they go through an interstitial
view. Because we apply URL rewrite logic globally through middleware
that processes all Django requests [0], using these tags in templates is
redundant and can be misleading; for example, using unsigned_redirect in
a template gets converted to a signed_redirect in the middleware.

It turns out that unsigned_redirect isn't really used: anywhere that we
redirect through the interstitial, we want to sign the URL so that we
only redirect to intended destinations. So this commit also removes that
method.

[0] https://github.com/cfpb/cfgov-refresh/blob/d48b4a5af6e928bf35fe982f7eeb96793c497e46/cfgov/cfgov/settings/base.py#L118