Validate our static assets in CI
We've had a `validate-assets` tox environment for ages that collects our assets and then runs Django's `collectstatic` on them. Because we use `ManifestStaticFileStorage` in production, if static assets are missing, this `collectstatic` will fail.

Because our webfonts are not public, we have not been able to run this in CI.  I've pulled in a workaround for this (https://github.com/cfpb/cfgov-refresh/pull/5110/files#diff-300249821afbbfcb945e6fbe141d0bb8R32-R55) created by @chosak in another context in #5110.