Fix wellbeing server errors caused by translations
After the deployment of #5939, we started getting server errors on the
Spanish results page for the wellbeing questionnaire.

The server errors are caused by a KeyError originating in the
wellbeing view, which was translating keys in a data dict.
It turns out we don't need to translate those keys, because the template
can handle any needed translations.

Testing
Locally, this page should raise a 500 error:

http://localhost:8000/consumer-tools/financial-well-being/results/es/

If you pull in this branch, it should render again.

Note
There remains an odd translation failure for the "Age" comparison button, which should read
"Edad" on the Spanish results page, but instead is blank.

The translation in the cfgov/locale/es/LC_MESSAGES/django.po file is correct.
I tried retyping the entry and regenerating the django.mo file, to no effect.
This is not related to the current bug, and should not hold up its resolution.