Properly validate translations
The existing validate-translations tox environment doesn't properly
validate translations, because it tests the wrong path.

The tox config tries to test files under cfgov/locale, but it was using
a changedir setting that changed the working directory to cfgov. So in
practice it wasn't actually testing any files. Currently master does
generate some changes in translation files, but these were being missed
because of this misconfiguration.

This change modifies the tested path from cfgov/locale to just locale,
so that the right files are tested.