Use gettext_noop and a list of tuples in the FWB ResultsView
This change restores gettext around the FWB groups and changes those groups from `OrderedDict`s to a list of tuples. This should avoid any key-errors with translated strings when looping over `group_means` and still allow those strings to be picked up by gettext.

It also replaces `gettext_lazy` in the FWB views with `gettext_noop`, since we want to mark the strings for translation but do not want them actually translated when the view is evaluated (that happens in the template).

I've also added gettext to the Dockerfile so that it's available in local development without having to yum install it.