Install wagtail-autocomplete from fork
We need to temporarily install a cfpb fork of wagtail-autocomplete to
facilitate editing of Ask CFPB content while the project's origin repo,
works on [a PR](https://github.com/wagtail/wagtail-autocomplete/pull/77) we pushed to them.

The PR has been approved, but it may take some time to get it merged and for
a new release to be issued to PyPI.

This is a temporary fix for https://github.local/CFGOV/platform/issues/3862

Testing

1. In current master, open http://localhost:8000/admin/pages/5804/edit/ and note the server error(s) generated by the autocomplete API calls. They will show up in your local server window and in the network pane of dev tools.
Also note that the related-questions and redirect-to-page fields won't work.
2. Pull in this branch, install libraries with `pip install -r requirements/local.txt`
This will install the forked version.
3. Note that the server errors go away when editing an Ask CFPB page, and the
autocomplete panels now function.