Add wagtail-placeholder-images library for local development
Currently, images that are uploaded to the production site go to an S3 
bucket. As a result, when loading a database dump from the production 
database for local development, most of the images do not show up on the 
page. This handy-dandy new pacakge from our friend @rinti replaces those 
missing images with placeholders from a service of our choice! 
(Currently using placedog.net, which was the suggested service in the 
package's README.)