Only check live EventPages for archiving; append date to duplicate slugs
Drafts that got archived were causing errors when, later, pages with the 
same slug were attempted to be archived again. This update only checks 
for live pages to archive (and only EventPages now), and if it finds a 
duplicate slug, it will append the event's start date to the slug before 
moving it.