Improve mega menu submenu highlighting
We want a mega menu submenu to be highlighted (i.e. appear visually
selected) when users are on a child page of that submenu.

The current logic assumes that we can determine this by comparing URL
paths; for example, if a submenu contains a link to /foo/, and the user
is on /foo/bar/, then we should highlight the submenu.

For increased flexibility, we also want to support the case when menu
items point to links with query strings, for example /foo/?something. In
this case, we want a page like /foo/bar/ to also mark the submenu as
selected. This commit alters the logic so that we strip off the menu
link query string before comparing it to the request path.

Note that this logic only applies to the top-level highlighting of
submenus (for example Consumer Tools on the current site). This doesn't
alter how the specific menu links themselves are highlighted.