Replaces the ElasticSearch 2.3 and 2.4 images with a single 2.3.5 image
This replaces the 2 separate ES 2.3 / 2.4 images, used in docker-compose.yml and docker-stack.yml, with a single image.

This image is based on CentOS7 and downloads and installs ES 2.3.5 from Elastic. It ensures the same image is used in both local development and in a docker cluster. We hypothesize this _may_ help with ES flakiness in the docker cluster, though that is admittedly a long shot.

Addresses APPOPS-1201