Replaces the ElasticSearch 2.3 and 2.4 images with a single 2.3.5 image (#6014)
* Replaces the ElasticSearch 2.3 and 2.4 images with a single 2.3.5 image

This replaces the 2 separate ES 2.3 / 2.4 images, used in docker-compose.yml and docker-stack.yml, with a single image.

This image is based on CentOS7 and downloads and installs ES 2.3.5 from Elastic. It ensures the same image is used in both local development and in a docker cluster. We hypothesize this _may_ help with ES flakiness in the docker cluster, though that is admittedly a long shot.

Addresses APPOPS-1201

* Add a restart policy for elasticsearch

We discussed as a team adding restart policies for the images. I figure it's OK to add one in this PR since it's addressing the general problem of ES flakiness

* Tweak restart_policy to match Alex's

I noticed that Alex pushed restart_policy changes for all images; this mirrors his change for ES