Add script for validating redirects
This new script allows for validation of a set of redirects as defined
in a CSV or XLSX file. The first line of the file will be ignored as
headers. The "from" and "to" paths must live in the 6th and 7th column
of the input file, respectively.

To run against www.consumerfinance.gov, using an input XLSX:

cfgov/scripts/validate_redirects.py redirects.xlsx

To run against localhost, using an input CSV:

cfgov/scripts/validate_redirects.py redirects.csv http://localhost:8000

To test, run without one of the redirects and observe that the check
fails. Now add the redirect and note that the check passes as long as it
redirects to the correct path.