Add optional to/from archive fields for filterable lists
This change adds two rich text fields for providing text and links to take users to the archive route on filterable lists, and return them.

`archive_link_text` will display on the filterable list index if it is provided and if the filterable list has archived content.

`return_from_archive_link_text` will display on the filterable list's archive route.