Add temporary script to archive content older than 2 years
This change adds a script that will take a filterable list page URL path as its argument and then archive the pages within that filterable list queryset if the page's `last_published_at` date is more than two years from when the script runs.

This is based on the scripts for the IA Refresh by @schbetsy.