Remove option to export page metadata as an .xlsx file
The XLSX download option was error-prone for page metadata reports,
possibly because of their size. The XLSX option required that the data be
held in memory before output, whereas the CSV option allows
for a more efficient streaming process.

In any case, not much is lost by removing the option, because the CSV can be
opened in Excel, which then saves the result as an .xlsx file by default.