Fix the nmdb CSV export by calling getvalue()
Not sure why this cropped up now, but along the way we lost the ability
to feed a bare file object to boto3 for uploading to S3. Boto creates and
md5 hash of the content and expects bytes in making the hash, and
our previous code delivered unicode.

The solution is to pass the bytes instead of the file object to boto,
by calling `getvalue()` on the file object.