Ensure that the paginator is always getting a valid page
This changes moves where we get the page number from the search form up into the same block where we ensure the form is valid.

If the form was not valid, we were still trying to get the `page` from the form's cleaned_data, and that could fail. This fixes that issue.