Convert is_archived field to a choice field with never archive option
We will soon be implementing an automatic nightly archiving script to 
archive anything that was published more than two years ago. As we may 
want exceptions to this rule, we needed a way to tell this script not to 
archive certain pages.

This commit converts `is_archived` from a boolean field into a CharField 
with defined choices, one of which is `'never'`, which, if chosen, will 
tell the forthcoming script to not archive that page.