Replace bulk archive script with management command
This change replaces the bulk archive script with a management command that can take multiple paths and can archive based on year, month, and day arguments based on either first or last published date.

By default this command will set pages to archived if their `first_published_at` date is 2 years or more from when the script is run.

See #6041 for more context on the bulk archive script.