Fix AnswerPage description auto-truncation when RTF does not come first
We started getting 500 errors on AnswerPages with empty short_answer 
fields and VideoPlayer blocks placed first in the answer_content 
StreamField.

This fixes that by cleaning up AnswerPage's description-setting code to 
piggyback on/override the method recently put in place for CFGOVPages as 
a whole.

It also makes the `truncate_preview` function a more clearly generic 
`truncate_by_words_and_chars` utility function. A future commit can pull 
that out and make use of it on the aforementioned CFGOVPage 
`get_meta_description` method.