Expose undergraduate programs in the paying-for-college API
The newish [financial-path tool](https://www.consumerfinance.gov/paying-for-college/your-financial-path-to-graduation/) is ready to start helping undergraduate users. The initial launch was aimed solely at
graduate programs, so undergrad programs were left out of the API.

This just opens the gate for the 'undergrad' portion of the schools' programCodes endpoint.

Testing
It helps to have a json viewing extension installed in your browser.
With a recent version of production data, check out the branch and visit the local endpoint for Texas Tech (ID 229115):
<http://localhost:8000/paying-for-college2/understanding-your-financial-aid-offer/api/school/229115/>

Note that the programCodes endpoint lists both graduate and undergrad programs.

Visit the live Texas Tech endpoint, and note that programCodes[undergrad] is an empty list.
<https://www.consumerfinance.gov/paying-for-college2/understanding-your-financial-aid-offer/api/school/229115/>