Remove redundant ArticlePage search field
The ask_cfpb.ArticlePage model defines a custom search field for the
'title' field, but this field is already included in the base
Page.search_fields:

https://github.com/wagtail/wagtail/blob/cc13aecd7d25163a5deac60c9aea6df3e0a45e1b/wagtail/core/models.py#L494

Redefining the field like this causes problems when trying to use
Wagtail search with an Elasticsearch backend, because it's not possible
to change a mapping once it has been created. See this related Wagtail
issue:

https://github.com/wagtail/wagtail/issues/4717

I do note that the base Page definition of the index is slightly
different:

index.SearchField('title', partial_match=True, boost=2)

as opposed to the simpler form used here:

index.SearchField('title')

The git blame doesn't indicate that this change was deliberate in the PR
that created this page type:

https://github.com/cfpb/consumerfinance.gov/pull/5009

The only place this might impact the live site is in our "external link
search", as Ask CFPB uses its own, separate Elasticsearch logic. Thus I
am inclined to make this change unless there is a specific reason for
this redefinition.