Add server indicator banner
In part to assuage potential concerns about removing the custom title 
from the login page, this adds a little banner to the top left of the 
admin calling attention to what hostname you're on. (It goes the full 
width of the screen on smaller viewports.)

At the moment, the banner is always the same color, but it might be nice 
to have it be color-coded for our common environments.