Use Python pathlib instead of third-party Unipath
This commit deprecates use of the third-party Unipath package and
instead uses the built-in Python pathlib package.

Unipath is now in maintenance mode [0]. We used it previously with
Python 2. Python 3.4 introduced pathlib [1] which has the same
functionality for our purposes.

We currently use Unipath for path-handling logic, but pathlib.Path
can do the same thing. The default Django project settings file uses
pathlib in this way [2].

There are no other uses of pathlib in any of the cf.gov satellite apps.
[3]

[0] https://github.com/mikeorr/Unipath
[1] https://docs.python.org/3/library/pathlib.html
[2] https://github.com/django/django/blob/45f4282149e13a2c1548a579f60d098e397a33d7/django/conf/project_template/project_name/settings.py-tpl#L13-L16
[3] https://github.com/search?q=user%3Acfpb+unipath&type=code