Fix: don't render empty info unit heading blocks
Currently if an info unit contains an empty heading block (meaning empty
text), an empty <div class="m-info-unit_heading-text"><h3></h3></div>
gets rendered on the page.

If a heading has empty text, it should not be rendered as part of an
info unit. This commit removes the branch of logic that does this.

To test, visit a page with an empty heading, for example in production:

http://localhost:8000/consumer-tools/educator-tools/adult-financial-education/

Alternatively, create a new e.g. Sublanding page with an info unit
group, then add an info unit with an empty heading.

Fixes internal platform#3854.