Attempt to fix flapping CCT chart button test
As originally written, the test couldn't guarantee that the first 3y 
button on the page would be the first one to load, resulting in the 
get/click sometimes happening on a later button than the assertion was 
then checking. This solves it by just getting the first one that loads, 
clicking it, and then immediately testing that that object has received 
the approriate class.