Override the TreeModelAdmin URL Helper for Regs breadcrumbs
This change overrides the TreeModelAdmin URL Helper for Regulations breadcrumbs. in #5703 we ensured that the regs models' `__str__` methods included their parents, so that one could see what regulation and version a particular section object belongs to in the model drop-downs and anywhere else Django renders the model to string.

Unfortunately the `__str__` is the same mechanism TreeModelAdmin uses to generate the breadcrumbs at the top, which get truncated at a particular length. This seems reasonable.

So to eliminate the duplicity in the breadcrumbs, this change overrides the TreeModelAdmin URL helper class to remove the parent's string if the URL helper is generating a breadcrumb for a specific model and a parent instance is given.