Fix the docker attach line in docs
This change fixes the line in our docs for the `docker attach` command when you need to interact with a `pdb` session. I think when we renamed this repo from `cfgov-refresh` to `consumerfinance.gov`, we mistakenly broke this command. It should not have the dot in the container name; it should be `consumerfinancegov_python_1`.