Improve keyboard navigation of VideoPlayer
This commit improves keyboard navigation of the VideoPlayer module by
changing which elements are assigned tabstops.

This fixes https://github.com/cfpb/design-system/issues/1124 which
reports that tabbing through the video player pauses on invisible
elements. This is because currently both the close button and video
iframe can be tabbed to even when they are invisible because the
video is not playing. An example live page where this behavior can be
seen is
https://www.consumerfinance.gov/consumer-tools/managing-someone-elses-money/.

This commit adds tabindex="-1" to both the iframe and close button so
that, in the initial page load state, neither of these invisible
elements can be tabbed to.

The JS code is altered here so that, when the video is played, the
tabindex attribute is removed, thus making these elements tabbable. The
play button also then gets tabindex="-1", as it gets hidden when the
video is played. Additionally, the "Close video" now gets the focus when
this happens (and also a new aria-label).

When the video is stopped, the JS code now reverses the process, and the
focus is set to the play button.

To test, use VoiceOver on the Mac (for example as documented at
https://www.imore.com/how-enable-voiceover-mac) and tab through the
page, noting that only appropriate elements can be selected. Use the
space bar to open/close the video. Alternatively or additionally, you
can also add a watch expression for document.activeElement in Chrome
developer tools, and observe how it changes as you tab.

One potential followup to this work would be to consider the placement
of the close button relative to the video controls. Currently the
module HTML puts both the play and close buttons AFTER the video iframe.
This means that someone using keyboard navigation would have to tab "in
reverse" to get to the video controls, where they might expect these to
come after play/close. Fixing this would require moving the buttons
before the iframe, and I'm not sure what else this might impact.