Improve Info Unit image alt text logic (#6202)
This commit modifies how Info Unit image alt text is set. With this
change, the following logic is used:

- If the image is not a link, use its alt text. This might be the alt
text configured on the Info Unit, or, as a fallback, the alt text set on
the image when added to Wagtail.
- If the image is a link, and the Info Unit has a heading, don't set any
alt text. The link will already include the heading text, and we don't
want there to be duplicate text in the link.
- If the image is a link, and the Info Unit doesn't have a heading, make
the alt text equal to the link text. If we used image-specific alt text
here, then we'd have two links going to the same place, but with
different text. Using the link text means that both links are the same.
This is consistent with how the carousel links work on the homepage mega
menu.

This is followup work from PR 6168.