Move video iframe below the video's actions container
Having the video player iframe above the video's actions makes it difficult
for keyboard navigators to control the video. They have to start the video
and then reverse-tab up to the youtube iframe to access it.

This change puts the video after the actions container so that you can play
it and then tab into the youtube iframe.