Change Featured Content body to plain text (#6217)
The Featured Content Module (FCM) is documented in the design system:

https://cfpb.github.io/design-system/patterns/featured-content-module

Its Wagtail block implementation is
v1.atomic_elements.organisms.FeaturedContent.

The design documentation mentions "descriptive body copy text" and that

> Descriptions should be action-oriented and let the user know what they
will get out of the featured content and how it’s related to the main
page content, not just what it is. Descriptions should be 160 characters
or less (including spaces).

This description and the documented examples suggest that the body text
should be plain text and thus should not contain links, images, or text
formatting). But in code this field is implemented as a Wagtail
RichTextBlock, which allows for all of these things:

https://docs.wagtail.io/en/stable/topics/streamfield.html#richtextblock

This commit modifies the FCM body field to instead by a CharBlock:

https://docs.wagtail.io/en/stable/topics/streamfield.html#charblock

This allows for only a single line of plain text.