Stop server errors when copying pages (#6231)
Currently, an attempt to copy a page in Wagtail copies the page and its tags,
but then the admin API raises a server error when it tries to handle tags.
If we exclude tags from admin operations, the errors cease, and page copying
can complete as expected.

This sidesteps a broader question about whether author tags should get copied
to a new page. For now, tags and authors will still get copied to the new page.

Testing

- In the local page, pick a page with tags and authors. This Newsroom Page works:
    <http://localhost:8000/admin/pages/3395/>
- From the "MORE" button, choose "Copy," change the title and slug, and click on
COPY THIS PAGE.

You should not get an error, and if you search for your new title, you should find
the new page, and the previous tag and author should be applied.