Remove unnecessary use of FilterableListForm
The SublandingPage class has a `get_browsefilterable_posts` method that
is used to generate a list of posts to display if the page has a
PostPreviewSnapshot module.

There are only two live pages on the site that use this functionality
now:

https://www.consumerfinance.gov/compliance/amicus/
https://www.consumerfinance.gov/rules-policy/notice-opportunities-comment/

As part of broader work on filters, this commit removes the use of
FilterableListForm here. Invoking the form does nothing because there
are no form filters being specified. Removing this use of FLF will make
it easier to make changes to that code.