Reintroduce homepage modifier for one-off hero fixes
Before restructuring the homepage for the carousel, we used a modifier
to tweak the hero. We are now reverting to the former hero layout so I'm
bringing back the home.less modifier file.

Fortunately, we have a snazzy new jumbo hero (see https://github.com/cfpb/consumerfinance.gov/pull/5614)
so very few of the tweaks need to be revived.

Note: We are redesigning the homepage so all this code will be replaced
in a few months. Consider this an interim fix.