Restore two Ask CFPB Spanish translations
1. "Read answer" on the landing page

This translation was automatically commented out in the translation file 
for unknown reasons. I manually restored it.

2. "Was this answer helpful to you?" on the feedback module

It appears that this translation was automatically commented out due to 
the text being un-hard-coded in our templates and moved to a field in 
Wagtail. I manually restored it.

I'm not sure if either will get commented again when someone else runs 
`makemessages`. We really need to figure out a way to prevent that from 
happening in the future.