Use Elasticsearch for FilterableLists (#6262)
* Initial iteration on ES Search

* Use queryset instead of mapping fields

* Funcitoning search for events and enforcement actions

* Fix issues with Activity Log

* Linter fixes

* Update cfgov/v1/atomic_elements/organisms.py

Co-authored-by: Andy Chosak <andy.chosak@cfpb.gov>

* Tests passing

* Cleanup document

* fix isort issues

* Update backend workflow to add elasticsearch service

* Create index where it is initially used as well

* linter fixes

* Remove time.sleep

* Resolve merge issues

* Cleanup form logic

* Block on Elasticsearch updates during tests (#6270)

This commit adds a new TestCase mixin for use during tests of
Elasticsearch using django-elasticsearch-dsl. By default, updates to
Elasticsearch are not blocking. This can lead to some unpredictability
if we do tests like this:

1. Put some data in the database
2. Index this data into Elasticsearch (e.g. by calling `manage.py
search_index`)
3. Do some tests that query Elasticsearch

This is because, even though the code in step 2 may return, there's no
guarantee (by default) that the Elasticsearch index has finished and
will be reflected in queries.

ES does provide the ability to force this, by using `refresh='wait_for'`
[0], but unfortunately currently dj-es-dsl does not provide a way to
specify this option when updating.

This command adds a TestCase mixin that patches the dj-es-dsl update
command to add this parameter, forcing updates to work as if they are
blocking.

It also does a bit of cleanup/reorganization of the FilterableListForm
tests, using the standard Django convention of calling `form.is_valid`
to do cleaning, instead of manually calling `form.full_clean`.

[0] https://www.elastic.co/guide/en/elasticsearch/reference/master/docs-refresh.html

* Tests passing, linter passing, we're almost there

* Add tests for documents.py

* isort fixes

* filterable list mixins tests

* Refactor document test file for better coverage

* Fix import

* Add tests to cover old functionality of get_page_set

* Initial iteration on ES Search

* Use queryset instead of mapping fields

* Funcitoning search for events and enforcement actions

* Fix issues with Activity Log

* Linter fixes

* Tests passing

* Update cfgov/v1/atomic_elements/organisms.py

Co-authored-by: Andy Chosak <andy.chosak@cfpb.gov>

* Cleanup document

* fix isort issues

* Update backend workflow to add elasticsearch service

* Create index where it is initially used as well

* linter fixes

* Remove time.sleep

* Resolve merge issues

* Cleanup form logic

* Block on Elasticsearch updates during tests (#6270)

This commit adds a new TestCase mixin for use during tests of
Elasticsearch using django-elasticsearch-dsl. By default, updates to
Elasticsearch are not blocking. This can lead to some unpredictability
if we do tests like this:

1. Put some data in the database
2. Index this data into Elasticsearch (e.g. by calling `manage.py
search_index`)
3. Do some tests that query Elasticsearch

This is because, even though the code in step 2 may return, there's no
guarantee (by default) that the Elasticsearch index has finished and
will be reflected in queries.

ES does provide the ability to force this, by using `refresh='wait_for'`
[0], but unfortunately currently dj-es-dsl does not provide a way to
specify this option when updating.

This command adds a TestCase mixin that patches the dj-es-dsl update
command to add this parameter, forcing updates to work as if they are
blocking.

It also does a bit of cleanup/reorganization of the FilterableListForm
tests, using the standard Django convention of calling `form.is_valid`
to do cleaning, instead of manually calling `form.full_clean`.

[0] https://www.elastic.co/guide/en/elasticsearch/reference/master/docs-refresh.html

* Tests passing, linter passing, we're almost there

* Add tests for documents.py

* isort fixes

* filterable list mixins tests

* Refactor document test file for better coverage

* Fix import

* Add tests to cover old functionality of get_page_set

* Elasticsearch-based filters suggestions (#6272)

* Test cleanup, rebuild_elasticsearch_index helper

* Remove unused content field for now

* Use PageQuerySet.specific() to optimize indexing

* Add tests for filterable_categories

* Linter fixes

* Test all filterable forms

* Swap one test class to use autosync to get full coverage

* edge ngram tokenizer for title and refactor name

Co-authored-by: Alex Martin <alexander.martin@cfpb.gov>
Co-authored-by: Andy Chosak <andy.chosak@cfpb.gov>