Use Django forms for Ask CFPB input
This change will use Django forms as the place where we clean and validate our Ask CFPB search inputs, rather than doing it directly from `request.GET` in the views.

It also validates the legacy facets against the expected pattern and adds unittests for the legacy facet redirect.

It also switches to f-strings for the string formatting of the URL paths/names.