Enable git+https URLs in our requirements files
In order to use `git+https://` URLs in our requirements files, we need git to be able to successfully clone. Prior to this change, git could not clone in our build container, failing with the error: `fatal: unable to look up current user in the passwd file: no such user`.

This changes sets `GIT_COMMITTER_NAME` and `GIT_COMMITTER_EMAIL` in our build script. git requires either the user its running as to exist in the `passwd` file *or* for these two variables to be set. By setting them here, we satisfy git's requirements and are able to include `git+https` URLs.

git 2.6.5+ apparently removes this requirement.