Add reusable text to legacy blog pages
Occasionally the CFPB mission statement in the blurb at the end of
press releases changes. To make it easier to manage those changes, we
want to pull that blurb out of the raw HTML content of legacy blog
and newsroom pages and put it in a reusable text snippet. That way
the blurb just needs to be updated in one place any time it changes.