Remove the POST from external site interstitial view
This change removes POST support from the external site interstitial view.

The POST was being used to POST URLs that didn't necessarily match the URL in the GET. This was causing issues in production.

The POST also seems unnecessary, as it adds another step through our servers before the redirect, when we can simply link directly to the external site from the button.

This change does require adding /external-site/` to the parse links exception list, because without it, the user ends up in an endless loop of being sent to /external-site/ when `ParseLinksMiddleware` parses the rendered interstitial page.