Enable Apache server status endpoint
This change enables Apache's mod_status and adds a server status endpoint that sits behind a check of a new environment variable, `APACHE_STATUS_ALLOW_FROM`. If that variable is defined, the endpoint is added, and the `Allow from` for access to it is populated from its contents. By default the endpoint will `Deny from all`.

In the dev prod-like Docker compose file I've set this to "all" to enable use and debugging locally.