Add middleware to explicitly allow CSRF on paths
Because the Django CSRF middleware processes the view for each request, the easiest way for us to exempt whole parts of the same from CSRF requirements is to use custom middleware to do just that, per https://github.com/wagtail/wagtail/issues/3066.

This middleware will check to see if a request's path matches a set of configurable paths in settings, and apply CSRF requirements to any paths that do match.