Simple Chart: Show staging data on draft pages
If the page is being served via wagtail-sharing (i.e. the request's
served_by_wagtail_sharing attribute is set to True), replace any
instances of 'production' in the data_source field with 'staging'.
If the page is not being served via wagtail-sharing, do not modify the
data-source field.
This allows us to provide a staged version of a page with different data
than is on production. We're assuming that we can have production and
staging data sources that exactly parallel each other in URL structure,
but replacing 'staging' with 'production'.
We're initially building this with CCPI in mind, but hope this feature
is general enough to be used for other charts.